Added JCusolver samples
The JCusolver samples that are added here are direct
ports of the original NVIDIA samples. Cleaned-up
versions of these samples might be added later.

The samples use the "MatrixMarketReader" utility
library, which was deployed into a local Maven
repository.

The samples read the MatrixMarket files that are
shipped together with the NVIDIA samples, and which
are also contained in the resources directory.